Refactor popup component
