+++ src/pages/popup/config.ts
@@ -1,0 +1,28 @@
+import __ from '../shared/i18n';
+
+export const introSteps = [
+  {
+    element: '.header',
+    intro: __('introHeader'),
+  },
+  {
+    element: '#vu_meter',
+    intro: __('introVuMeter'),
+  },
+  {
+    element: '#enabled',
+    intro: __('introEnabled'),
+  },
+  {
+    element: '#speed-settings',
+    intro: __('introSpeed'),
+  },
+  {
+    element: '#silence_threshold',
+    intro: __('introSilenceThreshold'),
+  },
+  {
+    element: '#mute_silence',
+    intro: __('introMuteSilence'),
+  },
+];
